Add Scanner option to expand variations (part 2 of 5)
This commit is part of a series which adds a feature to the Scanner to
expand variations into individual Game instances during parsing. This
commit specifically adds Game.Split(), the core functionality
introduced by this series. Given a Game with a main line and
possibly many variations, Game.Split() will return a slice of Games, 1
per variation, which each individually have only a single main line.
This makes it convenient for callers who can then utilize Game.Moves()
and other methods which only work on the main line.