Add Scanner option to expand variations (part 3 of 5)
This commit is part of a series which adds a feature to the Scanner to
expand variations into individual Game instances during parsing. This
commit specifically adds Scanner.ParseNext(). Scanner.ParseNext() is a
higher level convenience iterator combining the functionality of
ScanGame(), TokenizeGame(), NewParser(), and Parse(). This will later
be extended to support expanded variations in part 4.