Add Scanner option to expand variations (part 4 of 5)
This commit is part of a series which adds a feature to the Scanner to
expand variations into individual Game instances during parsing. This
commit specifically adds ScannerOption, and the first ScannerOption
WithExpandVariations(). Additionally this commit implements the option
by taking advantage of part 2's Game.Split() function within
Scanner.ParseNext() and caches the result in Scanner.nextParsedGames
to be utilized in subsequent invocations.