(bugfix) Game.String() should not assume the first move is move 1
When a Game is constructed from a .pgn containing a FEN, the parsed
moves may not begin at move 1. Similarly, the player to move cannot be
assumed to be white. This commit fixes both by using the rootMove's
position to determine both the initial move number to output and the turn.