Bug fixes for working with .PGNs that contain FEN position tags. (#58)
* (bugfix) Lexer.NextToken() is missing * result token

* (bugfix) parseVariation() should not assume starting position

When parsing a game which includes both a FEN position tag and
variations, parseVariation() should use the rootMove's position rather
than assume the starting position.

* (bugfix) Game.String() should not assume the first move is move 1

When a Game is constructed from a .pgn containing a FEN, the parsed
moves may not begin at move 1. Similarly, the player to move cannot be
assumed to be white. This commit fixes both by using the rootMove's
position to determine both the initial move number to output and the turn.

---------

Co-authored-by: Corentin Giaufer Saubert <43623834+CorentinGS@users.noreply.github.com>