(bugfix) Game.PushNotationMove() fails to push for PGN games
When a Game is initialized via PGN(), subsequent moves appended to the
original via Game.PushNotationMove() are not being appended. (The
game's position gets updated, but the move is missing). The root cause
is Parser.Parse() fails to initialize Game.currentMove to the last
move of the PGN, so Game.addNewMove() was not adding the move as expected.
Added a new test to catch this as well as augmented an existing PGN
read test.